requirements.txt
pandas==2.3.3
numpy==2.0.2
requests==2.32.5
PyYAML==6.0.3
python-telegram-bot==13.15
ta==0.11.0
tzlocal==5.3.1
streamlit==1.39.0
